POIs can be filtered by category now, bug fixes, refactoring

+++ core/domain/src/main/java/com/bytecause/domain/usecase/UpdateVesselsUseCase.kt
@@ -3,41 +3,42 @@ package com.bytecause.domain.usecase
 import com.bytecause.domain.abstractions.VesselsDatabaseRepository
 import com.bytecause.domain.abstractions.VesselsPositionsRemoteRepository
 import com.bytecause.domain.model.ApiResult
-import com.bytecause.domain.model.VesselModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.flow
 
-class FetchVesselsUseCase(
+class UpdateVesselsUseCase(
     private val vesselsDatabaseRepository: VesselsDatabaseRepository,
     private val vesselsPositionsRemoteRepository: VesselsPositionsRemoteRepository,
 ) {
-    operator fun invoke(): Flow<ApiResult<List<VesselModel>>> =
-        vesselsDatabaseRepository.loadAllVessels().map {
-            val result = when {
-                it.isEmpty() -> updateVesselsFromRemote()
-                shouldUpdateVessels() -> {
-                    vesselsDatabaseRepository.deleteAllVessels()
-                    updateVesselsFromRemote()
-                }
-
-                else -> ApiResult.Success(data = it)
+    operator fun invoke(): Flow<ApiResult<Unit>> = flow {
+        when {
+            shouldUpdateVessels() -> {
+                vesselsDatabaseRepository.deleteAllVessels()
+                val result = updateVesselsFromRemote()
+                emit(result)
             }
-            result
+
+            vesselsDatabaseRepository.isVesselDatabaseEmpty().firstOrNull() == true -> {
+                val result = updateVesselsFromRemote()
+                emit(result)
+            }
+
+            else -> emit(ApiResult.Success(Unit))
         }
+    }
 
     private suspend fun shouldUpdateVessels(): Boolean {
         return vesselsDatabaseRepository.shouldUpdateVesselDatabase(System.currentTimeMillis())
             .firstOrNull() == true
     }
 
-    private suspend fun updateVesselsFromRemote(): ApiResult<List<VesselModel>> {
+    private suspend fun updateVesselsFromRemote(): ApiResult<Unit> {
         return vesselsPositionsRemoteRepository.parseXml().let { result ->
             when {
                 result.exception == null && result.data != null -> {
                     vesselsDatabaseRepository.addAllVessels(result.data)
-                    val updatedVessels = vesselsDatabaseRepository.loadAllVessels().firstOrNull()
-                    ApiResult.Success(data = updatedVessels)
+                    ApiResult.Success(Unit)
                 }
 
                 else -> {
@@ -50,3 +51,4 @@ class FetchVesselsUseCase(
         }
     }
 }
+
